requirements.txt
alembic==1.12.0
asgiref==3.7.2
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
python-dotenv==1.0.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
typing_extensions==4.8.0
Werkzeug==2.3.7
